@@ -3,18 +3,14 @@ from pathlib import Path
 from typing import Any
 
 import beancount.loader
-from beancount.core.data import Note
-from beancount.ingest import importer  # type: ignore[import]
+from beancount.core.data import Balance, Note, Transaction
+from beangulp.importer import Importer  # type: ignore
 
 from starling_beancount import extractor
 
 
 def filt_notes(entry: Any, account_name: str) -> bool:
-    if (
-        isinstance(entry, Note)
-        and entry.account == account_name
-        and "bean-extract" in entry.comment
-    ):
+    if isinstance(entry, Note) and entry.account == account_name and "bean-extract" in entry.comment:
         return True
     else:
         return False
@@ -35,7 +31,7 @@ def last_date(bean_path: Path, account_name: str) -> date:
     return max_date
 
 
-class StarlingImporter(importer.ImporterProtocol):  # type: ignore[no-any-unimported]
+class StarlingImporter(Importer):  # type: ignore
     def __init__(
         self,
         config_path: Path,
@@ -54,19 +50,19 @@ class StarlingImporter(importer.ImporterProtocol):  # type: ignore[no-any-unimpo
     def name(self) -> str:
         return self.account_name
 
-    def identify(self, file: "FileMemo") -> bool:
-        return self.acc in file.name
+    def identify(self, filepath: str) -> bool:
+        return self.acc in filepath
 
-    def extract(self, file: str, existing_entries: Any = None) -> list:
+    def extract(self, filepath: str, existing_entries: Any = None) -> list[Transaction | Balance | Note]:
         since = last_date(self.bean_path, self.account_name) - timedelta(days=self.lag)
         res = extractor.extract(self.config_path, self.acc, self.token_path, since)
         return res
 
-    def file_account(self, file):
+    def file_account(self, filepath: str):
         return self.account_name
 
-    def file_name(self, file):
+    def file_name(self, filepath: str):
         return None
 
-    def file_date(self, file):
+    def file_date(self, filepath: str):
         return None
